+++ packages/nimble-components/src/table-column/enum-text/models/table-column-enum-text-validator.ts
@@ -1,4 +1,4 @@
-import type { Mapping } from '../../../mapping/base';
+import { MappingText } from '../../../mapping/text';
 import type { ColumnInternals } from '../../base/models/column-internals';
 import {
     TableColumnEnumBaseValidator,
@@ -13,14 +13,7 @@ const enumTextValidityFlagNames = [...enumBaseValidityFlagNames] as const;
 export class TableColumnEnumTextValidator extends TableColumnEnumBaseValidator<
     typeof enumTextValidityFlagNames
 > {
-    public constructor(
-        columnInternals: ColumnInternals<unknown>,
-        supportedMappingElements: readonly (typeof Mapping)[]
-    ) {
-        super(
-            columnInternals,
-            enumTextValidityFlagNames,
-            supportedMappingElements
-        );
+    public constructor(columnInternals: ColumnInternals<unknown>) {
+        super(columnInternals, enumTextValidityFlagNames, [MappingText]);
     }
 }