+++ packages/nimble-components/src/wafer-map/modules/wafer-map-update-tracker.ts
@@ -5,7 +5,7 @@ import { UpdateTracker } from '../../utilities/models/update-tracker';
 const trackedItems = [
     'canvasWidth',
     'canvasHeight',
-    'quadrant',
+    'originLocation',
     'dies',
     'maxCharacters',
     'colorScale',
@@ -31,7 +31,7 @@ export class WaferMapUpdateTracker extends UpdateTracker<typeof trackedItems> {
         return (
             this.isTracked('canvasWidth')
             || this.isTracked('canvasHeight')
-            || this.isTracked('quadrant')
+            || this.isTracked('originLocation')
             || this.isTracked('dies')
             || this.isTracked('maxCharacters')
             || this.isTracked('colorScale')
@@ -48,7 +48,7 @@ export class WaferMapUpdateTracker extends UpdateTracker<typeof trackedItems> {
     }
 
     public get requiresScalesUpdate(): boolean {
-        return this.isTracked('quadrant') || this.isTracked('dies');
+        return this.isTracked('originLocation') || this.isTracked('dies');
     }
 
     public get requiresLabelsFontSizeUpdate(): boolean {